+++ R/lsm_c_pladj.R
@@ -121,8 +121,10 @@ lsm_c_pladj_calc <- function(landscape) {
         landscape <- raster::as.matrix(landscape)
     }
 
-    landscape_padded <- pad_raster(landscape, pad_raster_value = -999,
-                                   pad_raster_cells = 1)[[1]]
+    landscape_padded <- pad_raster(landscape,
+                                   pad_raster_value = -999,
+                                   pad_raster_cells = 1,
+                                   return_raster = FALSE)[[1]]
 
     tb <- rcpp_get_coocurrence_matrix(landscape_padded,
                                       directions = as.matrix(4))