+++ R/lsm_l_pladj.R
@@ -123,7 +123,8 @@ lsm_l_pladj_calc <- function(landscape) {
 
     landscape_padded <- pad_raster(landscape,
                                    pad_raster_value = -999,
-                                   pad_raster_cells = 1)[[1]]
+                                   pad_raster_cells = 1,
+                                   return_raster = FALSE)[[1]]
 
     tb <- rcpp_get_coocurrence_matrix(landscape_padded,
                                       directions = as.matrix(4))